build scripts: Monster re-write: Using global exports and created simple 4 lines for variants and have 1 master build script using export vars to make things a whole lot easier

+++ build_master.sh
@@ -1,10 +1,16 @@
 #!/bin/sh
+export PLATFORM="AOSP"
+export MREV="JB4.2"
+export CURDATE=`date "+%m.%d.%Y"`
+export MUXEDNAMELONG="KT-SGS4-$MREV-$PLATFORM-$CARRIER-$CURDATE"
+export MUXEDNAMESHRT="KT-SGS4-$MREV-$PLATFORM-$CARRIER*"
+export KTVER="--$MUXEDNAMELONG--"
 export KERNELDIR=`readlink -f .`
 export PARENT_DIR=`readlink -f ..`
 export INITRAMFS_DEST=$KERNELDIR/kernel/usr/initramfs
-export INITRAMFS_SOURCE=`readlink -f ..`/Ramdisks/AOSP_VZW
-export CONFIG_AOSP_BUILD=y
-export PACKAGEDIR=$PARENT_DIR/Packages/AOSP
+export INITRAMFS_SOURCE=`readlink -f ..`/Ramdisks/$PLATFORM"_"$CARRIER
+export CONFIG_$PLATFORM_BUILD=y
+export PACKAGEDIR=$PARENT_DIR/Packages/$PLATFORM
 #Enable FIPS mode
 export USE_SEC_FIPS_MODE=true
 export ARCH=arm
@@ -43,12 +49,10 @@ rm $PACKAGEDIR/zImage
 rm arch/arm/boot/zImage
 
 echo "Make the kernel"
-make VARIANT_DEFCONFIG=jf_vzw_defconfig SELINUX_DEFCONFIG=jfselinux_defconfig SELINUX_LOG_DEFCONFIG=jfselinux_log_defconfig KT_jf_defconfig
+make VARIANT_DEFCONFIG=jf_$CARRIER"_defconfig" SELINUX_DEFCONFIG=jfselinux_defconfig SELINUX_LOG_DEFCONFIG=jfselinux_log_defconfig KT_jf_defconfig
 
-curdate=`date "+%m.%d.%Y"`
-ktver="--KT-SGS4-AOSP-JB-VZW-"$curdate"--"
-echo "Modding .config file - "$ktver
-sed -i 's,CONFIG_LOCALVERSION="-KT-SGS4",CONFIG_LOCALVERSION="'$ktver'",' .config
+echo "Modding .config file - "$KTVER
+sed -i 's,CONFIG_LOCALVERSION="-KT-SGS4",CONFIG_LOCALVERSION="'$KTVER'",' .config
 
 HOST_CHECK=`uname -n`
 if [ $HOST_CHECK = 'ktoonsez-VirtualBox' ] || [ $HOST_CHECK = 'task650-Underwear' ]; then
@@ -61,10 +65,9 @@ fi;
 
 echo "Copy modules to Package"
 cp -a $(find . -name *.ko -print |grep -v initramfs) $PACKAGEDIR/system/lib/modules/
-cp 00post-init.sh $PACKAGEDIR/system/etc/init.d/00post-init.sh
-cp enable-oc.sh $PACKAGEDIR/system/etc/init.d/enable-oc.sh
-cp /home/ktoonsez/workspace/com.ktoonsez.KTweaker.apk $PACKAGEDIR/system/app/com.ktoonsez.KTweaker.apk
-cp ../Ramdisks/libsqlite.so $PACKAGEDIR/system/lib/libsqlite.so
+if [ $ADD_KTWEAKER = 'Y' ]; then
+	cp /home/ktoonsez/workspace/com.ktoonsez.KTweaker.apk $PACKAGEDIR/system/app/com.ktoonsez.KTweaker.apk
+fi;
 
 if [ -e $KERNELDIR/arch/arm/boot/zImage ]; then
 	echo "Copy zImage to Package"
@@ -73,26 +76,27 @@ if [ -e $KERNELDIR/arch/arm/boot/zImage ]; then
 	echo "Make boot.img"
 	./mkbootfs $INITRAMFS_DEST | gzip > $PACKAGEDIR/ramdisk.gz
 	./mkbootimg --cmdline 'console = null androidboot.hardware=qcom user_debug=31 zcache' --kernel $PACKAGEDIR/zImage --ramdisk $PACKAGEDIR/ramdisk.gz --base 0x80200000 --pagesize 2048 --ramdisk_offset 0x02000000 --output $PACKAGEDIR/boot.img 
-	echo "Executing loki"
-	./loki_patch-linux-x86_64 boot abootvzw.img $PACKAGEDIR/boot.img $PACKAGEDIR/boot.lok
-	rm $PACKAGEDIR/boot.img
-	#cp loki_flash $PACKAGEDIR/loki_flash
+	if [ $EXEC_LOKI = 'Y' ]; then
+		echo "Executing loki"
+		./loki_patch-linux-x86_64 boot aboot$CARRIER.img $PACKAGEDIR/boot.img $PACKAGEDIR/boot.lok
+		rm $PACKAGEDIR/boot.img
+	fi;
 	cd $PACKAGEDIR
 	cp -R ../META-INF-SEC ./META-INF
 	rm ramdisk.gz
 	rm zImage
-	rm ../KT-SGS4-AOSP-JB-VZW*.zip
-	zip -r ../KT-SGS4-AOSP-JB-VZW-$curdate.zip .
+	rm ../$MUXEDNAMESHRT.zip
+	zip -r ../$MUXEDNAMELONG.zip .
 
 	time_end=$(date +%s.%N)
 	echo -e "${BLDYLW}Total time elapsed: ${TCTCLR}${TXTGRN}$(echo "($time_end - $time_start) / 60"|bc ) ${TXTYLW}minutes${TXTGRN} ($(echo "$time_end - $time_start"|bc ) ${TXTYLW}seconds) ${TXTCLR}"
 
-	FILENAME=../KT-SGS4-AOSP-JB-VZW-$curdate.zip
+	FILENAME=../$MUXEDNAMELONG.zip
 	FILESIZE=$(stat -c%s "$FILENAME")
 	echo "Size of $FILENAME = $FILESIZE bytes."
-	rm ../VZWversion.txt
-	exec >>../VZWversion.txt 2>&1
-	echo "KT-SGS4-AOSP-VZW-"$curdate", "$FILESIZE", http://ktoonsez.jonathanjsimon.com/sgs4/aosp/KT-SGS4-AOSP-JB-VZW-"$curdate".zip"
+	rm ../$MREV-$PLATFORM-$CARRIER"-version.txt"
+	exec >>../$MREV-$PLATFORM-$CARRIER"-version.txt" 2>&1
+	echo "$MUXEDNAMELONG, $FILESIZE, http://ktoonsez.jonathanjsimon.com/sgs4/$PLATFORM/$MUXEDNAMELONG.zip"
 	
 	cd $KERNELDIR
 else
